service/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.5
pydantic-settings==2.6.1
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
robotframework==7.1.1
robotframework-requests==0.10.0
alembic==1.14.0

